services/ingestion/requirements.txt
kafka-python
duckdb
json